+++ CHANGES.md
@@ -5,6 +5,7 @@ wavesurfer.js changelog
 - add additional type to `waveColor` and `progressColor` parameters to support linear gradients (#2345)
 - Regions plugin: increase region z-index to fix stacking inconsistencies (#2353)
 - add `hideCursor` option to hide the mouse cursor when hovering over the waveform (#2367)
+- Regions plugin: check `maxLength` before resizing region (#2374)
 
 5.2.0 (16.08.2021)
 ------------------